==> 2_parsing_table_from_mbc.ipynb <==
<여론M>에서 다운 받은 HTML에서 테이블 추출을 하기 위한 코드입니다.

import pandas as pd
from bs4 import BeautifulSoup

with open('2022_president.html', encoding="UTF-8") as file:
    soup = BeautifulSoup(file, 'html.parser')
tables = pd.read_html(str(soup), extract_links="body")
df = tables[0]

df

table = df.apply(lambda col: [v[0] if v[1] is None else v[1] for v in  col])
table = table.rename(columns=table.iloc[0]).drop(table.index[0])
table.to_csv("from_mbc.csv")

df
